# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ayoitson API
  version: 1.0.0
  description: |
    The Ayoitson REST API. Phase 4 redesign — uniform POST/PUT/DELETE
    semantics across every resource, structured error envelope on every
    response, X-API-Key auth on every endpoint except `GET /api/health` and
    `POST /api/auth/setup`. See `docs/SDLC.md` and the vault `[[API Redesign]]`
    note for the design memo.
  contact:
    name: SALM LLC / Ayoitson
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - apiKey: []
...
paths:
...
  /api/api-keys:
    get:
      tags: [auth]
      summary: List API keys
      description: Returns key metadata only. Raw keys and hashes are never returned.
      responses:
        '200':
          description: API key metadata
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ApiKeyMetadata' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      tags: [auth]
      summary: Create API key
      description: Creates a new operator/browser key. The raw key is returned once in this response.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApiKeyCreate' }
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedKey' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /api/api-keys/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    delete:
      tags: [auth]
      summary: Revoke API key
      responses:
        '200':
          description: API key revoke result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiKeyRevokeResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
  /api/channels:
    get:
      tags: [channels]
      summary: List channels
      responses:
        '200':
          description: Channel summaries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChannelSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [channels]
      summary: Create channel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChannelCreate' }
      responses:
        '201':
          description: Channel created
          content:
            application/json:
              schema:
                type: object
                required: [number]
                properties:
                  number: { type: integer }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/channels/numbers:
    get:
      tags: [channels]
      summary: List channel numbers
      responses:
        '200':
          description: Sorted channel numbers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ChannelSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/channels/{number}:
    parameters:
      - $ref: '#/components/parameters/ChannelNumber'
    get:
      tags: [channels]
      summary: Read channel
      parameters:
        - name: programless
          in: query
          required: false
          schema: { type: string, enum: ['true', 'false'] }
          description: When `true`, omit the programs array from the response.
      responses:
        '200':
          description: Channel object
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Channel' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [channels]
      summary: Update channel
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Channel' }
      responses:
        '200':
          description: Channel updated
          content:
            application/json:
              schema:
                type: object
                required: [number]
                properties:
                  number: { type: integer }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [channels]
      summary: Delete channel
      responses:
        '200':
          description: Channel removed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteAck' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/channels/{number}/programs:
    parameters:
      - $ref: '#/components/parameters/ChannelNumber'
    get:
      tags: [channels]
      summary: Stream channel programs
      responses:
        '200':
          description: Streamed JSON array of programs
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/channels/{number}/description:
    parameters:
      - $ref: '#/components/parameters/ChannelNumber'
    get:
      tags: [channels]
      summary: Channel descriptor
      responses:
        '200':
          description: Minimal channel descriptor
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChannelDescription' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/filler-lists/{id}/channels:
    parameters:
      - $ref: '#/components/parameters/ResourceId'
    get:
      tags: [filler-lists]
      summary: Channels using this filler list
      responses:
        '200':
          description: Channel list
          content:
            application/json:
              schema: { type: array, items: { type: object, additionalProperties: true } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/guide/channels/{number}:
    parameters:
      - $ref: '#/components/parameters/ChannelNumber'
      - name: dateFrom
        in: query
        required: true
        schema: { type: string, format: date-time }
      - name: dateTo
        in: query
        required: true
        schema: { type: string, format: date-time }
    get:
      tags: [guide]
      summary: Channel guide lineup
      responses:
        '200':
          description: Lineup
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Issued by the first-run setup flow (`scripts/first-run.js` or
        `POST /api/auth/setup`). Format: `ayo_<base64url>`. Verified server-
        side via argon2id against the `api_keys` SQLite table.
  parameters:
    ChannelNumber:
      name: number
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
    ResourceId:
      name: id
      in: path
      required: true
      schema: { type: string, minLength: 1 }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - PAYLOAD_TOO_LARGE
            - RATE_LIMITED
            - INTERNAL
            - UPSTREAM_ERROR
            - SERVICE_UNAVAILABLE
        message: { type: string }
        details:
          type: object
          additionalProperties: true
...
    CreatedKey:
      type: object
      required: [metadata, rawKey]
      properties:
        metadata: { $ref: '#/components/schemas/ApiKeyMetadata' }
        rawKey:
          type: string
          description: Raw API key value. Shown ONCE; never reissued.
    ApiKeyMetadata:
      type: object
      required: [id, name, scopes, createdAt, lastUsedAt, revokedAt]
      properties:
        id: { type: string }
        name: { type: string }
        scopes:
          type: array
          items: { type: string }
        createdAt: { type: string }
        lastUsedAt: { type: string, nullable: true }
        revokedAt: { type: string, nullable: true }
    ApiKeyCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    ApiKeyRevokeResult:
      type: object
      required: [revoked]
      properties:
        revoked: { type: boolean }
...
    ChannelSummary:
      type: object
      properties:
        number: { type: integer }
    Channel:
      type: object
      properties:
        number: { type: integer }
        name: { type: string }
        icon: { type: string }
        stealth: { type: boolean }
        programs:
          type: array
          items: { type: object, additionalProperties: true }
        fallback:
          type: array
          items: { type: object, additionalProperties: true }
      additionalProperties: true
    ChannelCreate:
      allOf:
        - $ref: '#/components/schemas/Channel'
        - type: object
          required: [number]
    ChannelDescription:
      type: object
      required: [number]
      properties:
        number: { type: integer }
        icon: { type: string }
        name: { type: string }
        stealth: { type: boolean }
...
    DeleteAck:
      type: object
      properties:
        deleted: { type: boolean }
        id: { type: string }
        number: { type: integer }
        name: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid X-API-Key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: Request body or parameter is invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ServiceUnavailable:
      description: Server starting up or shutting down
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }